Added terminal clear functionality. Minor Bug Fixes. Pushed test.cpp to main.cpp
